# docs/swagger.yaml
basePath: /beta
definitions:
  ErrorResponse:
    properties:
      message:
        example: example error message
        type: string
      results:
        example: 0
        type: integer
      success:
        example: false
        type: boolean
    type: object
...
  GetSingletonStatesArgs:
    properties:
      amount:
        example: 1
        format: int64
        type: integer
      coin_id:
        example: 625799464319c8703ac2d0664af98cf45b9b306f7dcf717b1070d170bb5916a9
        format: hex
        type: string
      header_hash:
        example: 796c33c3905150e649211fdd9ed42c7c418758c30c321271973a7c792a5bd403
        format: hex
        type: string
      height:
        example: 2174316
        format: int64
        type: integer
      inner_puzzle_hash:
        example: 6b665c0e059050f71a1c3e8a7d5b58e4e1d7abbd02d937e9b5ab5abfd7f8eaba
        format: hex
        type: string
      launcher_id:
        example: f4dd6f4ec490974f7eb98223748f47340a9e9363b4c2dccc1932cdbbc54d03fd
        format: hex
        type: string
      limit:
        example: 1337
        type: integer
      offset:
        example: 1
        type: integer
      order:
        example: desc
        type: string
      order_by:
        type: string
      parent_coin_id:
        example: e9676e8ce096c5be27dee2fbf2120054d206e4df2de9ef59c24a651d3c558c95
        format: hex
        type: string
      puzzle_hash:
        example: 0a5a9c760970ebcc094c6f9faa3d9730f066c7a8f7450841a94fc4fd59229bc2
        format: hex
        type: string
    type: object
  GetSingletonStatesResponse:
    properties:
      results:
        example: 1
        type: integer
      singleton_states:
        items:
          $ref: '#/definitions/SingletonState'
        type: array
      success:
        example: true
        type: boolean
    type: object
...
  NoAPIKeyResponse:
    properties:
      message:
        example: No API key provided.
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  SingletonState:
    properties:
      amount:
        example: 1
        format: int64
        type: integer
      coin_id:
        example: 625799464319c8703ac2d0664af98cf45b9b306f7dcf717b1070d170bb5916a9
        format: hex
        type: string
      header_hash:
        example: 796c33c3905150e649211fdd9ed42c7c418758c30c321271973a7c792a5bd403
        format: hex
        type: string
      height:
        example: 2174316
        format: int64
        type: integer
      inner_puzzle_hash:
        example: 6b665c0e059050f71a1c3e8a7d5b58e4e1d7abbd02d937e9b5ab5abfd7f8eaba
        format: hex
        type: string
      launcher_id:
        example: f4dd6f4ec490974f7eb98223748f47340a9e9363b4c2dccc1932cdbbc54d03fd
        format: hex
        type: string
      melted:
        example: false
        type: boolean
      parent_coin_id:
        example: e9676e8ce096c5be27dee2fbf2120054d206e4df2de9ef59c24a651d3c558c95
        format: hex
        type: string
      puzzle_hash:
        example: 0a5a9c760970ebcc094c6f9faa3d9730f066c7a8f7450841a94fc4fd59229bc2
        format: hex
        type: string
    type: object
...
host: https://kraken.fireacademy.io
info:
  contact:
    email: y@kuhi.to
    name: yakuhito
    url: https://twitter.com/yakuh1t0
  description: The Beta FireAcademy.io API
  license:
    name: MIT
    url: https://github.com/FireAcademy/beta/blob/master/LICENSE
  termsOfService: https://fireacademy.io/terms-and-conditions.txt
  title: Beta API
  version: "1.0"
paths:
  /{api-key}/get_singleton_states:
    post:
      consumes:
      - application/json
      description: |-
        Singletons are like souls - when a coin dies (gets spent), they might move on to a new coin or disappear (melt). Beta views these transitions as changes of state.
        This endpoint takes multiple optional arguments and returns the states that match the criteria. Most body parameters are self-explanatory: if they match a field of the returned struct model (e.g., 'height' or 'launcher_id'), they act as filters. Only states with the specified values will be returned.
        The first special parameter is 'limit' - by default, this function returns 100 results at most. Use this parameter to tweak this value.
        'order_by' can be 'coin_id', 'header_hash', 'height', 'parent_coin_id', 'puzzle_hash', 'amount', 'launcher_id', or 'inner_puzzle_hash'. The default ordering is ascending, but that can be changed by setting the 'order' parameter to 'desc'.
        If your query returns more than 100 results and you need all of them for some reasons, you can also use the 'offset' parameter.
        Note: Two singleton states CAN have the same coin id but a different 'melted' value - the primary key is a composite one: (coin_id, melted)
      parameters:
      - description: The arguments (see description for more details)
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/GetSingletonStatesArgs'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GetSingletonStatesResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/NoAPIKeyResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: returns singleton states for given conditions
      tags:
      - Singleton States
...
  /get_singleton_states:
    post:
      consumes:
      - application/json
      description: |-
        Singletons are like souls - when a coin dies (gets spent), they might move on to a new coin or disappear (melt). Beta views these transitions as changes of state.
        This endpoint takes multiple optional arguments and returns the states that match the criteria. Most body parameters are self-explanatory: if they match a field of the returned struct model (e.g., 'height' or 'launcher_id'), they act as filters. Only states with the specified values will be returned.
        The first special parameter is 'limit' - by default, this function returns 100 results at most. Use this parameter to tweak this value.
        'order_by' can be 'coin_id', 'header_hash', 'height', 'parent_coin_id', 'puzzle_hash', 'amount', 'launcher_id', or 'inner_puzzle_hash'. The default ordering is ascending, but that can be changed by setting the 'order' parameter to 'desc'.
        If your query returns more than 100 results and you need all of them for some reasons, you can also use the 'offset' parameter.
        Note: Two singleton states CAN have the same coin id but a different 'melted' value - the primary key is a composite one: (coin_id, melted)
      parameters:
      - description: The arguments (see description for more details)
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/GetSingletonStatesArgs'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GetSingletonStatesResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/NoAPIKeyResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: returns singleton states for given conditions
      tags:
      - Singleton States
...
securityDefinitions:
  ApiKeyAuth:
    description: The API key can also be included in the URL or as a GET parameter
      - please see [https://docs.fireacademy.io/developers/using-api-keys](https://docs.fireacademy.io/developers/using-api-keys)
      for more details.
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"